+++ hutool-core/src/main/java/cn/hutool/core/text/NamingCase.java
@@ -157,7 +157,7 @@ public class NamingCase {
 	 * 将连接符方式命名的字符串转换为驼峰式。如果转换前的下划线大写方式命名的字符串为空，则返回空字符串。
 	 *
 	 * @param name   转换前的自定义方式命名的字符串
-	 * @param symbol 连接符
+	 * @param symbol 原字符串中的连接符连接符
 	 * @return 转换后的驼峰式命名的字符串
 	 * @since 5.7.17
 	 */