+++ CHANGELOG.md
@@ -12,11 +12,13 @@
 * 【core   】     增加Table实现（issue#2179@Github）
 * 【core   】     增加UniqueKeySet（issue#I4WUWR@Gitee）
 * 【core   】     阿拉伯数字转换成中文对发票票面金额转换的扩展（pr#570@Gitee）
+* 【core   】     ArrayUtil增加replace方法（pr#570@Gitee）
 * 
 ### 🐞Bug修复
 * 【core   】     修复ObjectUtil.hasNull传入null返回true的问题（pr#555@Gitee）
 * 【core   】     修复NumberConverter对数字转换的问题（issue#I4WPF4@Gitee）
 * 【core   】     修复ReflectUtil.getMethods获取接口方法问题（issue#I4WUWR@Gitee）
+* 【core   】     修复NamingCase中大写转换问题（pr#572@Gitee）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.7.22 (2022-03-01)